Move to_json to util
